% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwm.R
\name{cwm}
\alias{cwm}
\alias{is.cwm}
\alias{[.cwm}
\alias{summary.cwm}
\alias{print.cwm}
\alias{extract}
\alias{extract.cwm}
\alias{range_cwm}
\title{Community weighted mean of species attributes}
\usage{
cwm(com, traits, wstand = FALSE)

is.cwm(object)

\method{[}{cwm}(object, i, j, drop = FALSE)

\method{summary}{cwm}(object, long = F, ...)

\method{print}{cwm}(x, ...)

extract(x, ...)

\method{extract}{cwm}(x, what = "traits", ...)

range_cwm(object, max = 9, recalculate = TRUE)
}
\arguments{
\item{com}{Matrix or data frame with species composition data (samples x species).}

\item{traits}{Vector, matrix or data frame with species attributes (species x attributes). This can be \code{numeric} or \code{factor}.}

\item{wstand}{Logical. Should be the values in \code{traits} weighted-standardized prior to calculation of CWM? Weights are column sums in community data matrix \code{com}.}

\item{object, x}{Object of the class \code{cwm}}

\item{i, j}{Subscripts of the matrix of the class "cwm" (rows and columns).}

\item{drop}{Argument in the subsetting function, but currently not implemented (\code{[.cwm} always return data frame with at least one variable).}

\item{long}{should summary return long output? (TRUE vs FALSE)}

\item{...}{Other arguments passed into \code{as.matrix}, \code{summary} or \code{print} function. Currently not supported.}

\item{what}{Attributes extracted from the object of class \code{cwm}; either \code{traits} for the matrix of species attributes (species x traits), or \code{com} for matrix of species composition (samples x species).}

\item{max}{upper bound of the rescaled values}

\item{recalculate}{should the mean values in \code{object} be recalculated?}
}
\value{
Object of class \code{"cwm"}, which has \code{print}, \code{summary}, \code{as.matrix} and \code{'[.'} methods. Object of \code{"cwm"} class contains the matrix of calculated weighted means of species attributes for each sample (sample x weighted mean) and three attributes: \code{com}, species x sample matrix from which the weighted mean was calculated, \code{traits}, species x attributes matrix with species attributes, and \code{wstand}, logical value specifying whether traits were weighted-standardized prior to calculation of CWM (weights are column sums of \code{com}). All weighted means of species attributes must be based on the same species x sample matrix with the same number of samples.
}
\description{
Function \code{cwm} calculates weighted mean of species attributes, using matrix of species composition and one or several species attributes.

#' @rdname cwm
#' @export
as.matrix.cwm <- function (x, ...)
{
  attr (x, 'com') <- NULL
  attr (x, 'traits') <- NULL
  attr (x, 'wstand') <- NULL
  x <- as.matrix (unclass (x))
  return (x)
}
}
\details{
Function \code{[.cwm]} is for extracting specified rows and columns from matrix of class \code{cwm}. As a side effect, resulting object will have concatenated the \code{com} and \code{traits} attributes to match the dimension of the resulting matrix. This function is only for extracting the parts of \code{cwm} object, not for replacing - the attempt to replace will work, but will break the inner structure of the object.

Generic function \code{extract} with argument \code{what} applied on object of class \code{"cwm"} extracts the original species composition matrix or species attribute matrix, respectively, which were used to calculate weighted means.

Function \code{range_cwm} changes the range of species attributes and recalculates the CWM values.
}
\examples{
# Calculation of weighted mean of species Ellenberg indicator values using dataset Vltava
data (vltava)
mean.eiv <- cwm (com = vltava$spe, traits = vltava$eiv)

summary (mean.eiv)

# Extracting values from the object of cwm class
mean.eiv[,1]
mean.eiv[1:10, 2:3]
}
\author{
David Zeleny (zeleny.david@gmail.com)
}
