% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfit_cwm.R
\name{envfit_cwm}
\alias{envfit_cwm}
\title{Fits community weighted mean of species attributes (CWM) onto an ordination}
\usage{
envfit_cwm(ord, cwm, wstand = attr(cwm, "wstand"), permutations = 999,
  choices = c(1, 2), display = "sites", w = weights(ord, display),
  na.rm = FALSE, test = "modified", adjustP = FALSE,
  p.adjust.method = "holm", p.adjust.n = NULL, ...)
}
\arguments{
\item{ord}{An ordination object or other structure from which the ordination scores can be extracted (including a data frame or matrix of scores).}

\item{cwm}{Community weighted means calculated by function \code{cwm}; object of the class \code{cwm}.}

\item{wstand}{Logical; should be the cwm calculated on weighted standardized species attributes? Defaults to the same setting as used in \code{cwm} function (FALSE).}

\item{permutations}{Number of permutations for the permutation test.}

\item{choices}{Axes to be plotted.}

\item{display}{In fitting functions these are ordinary site scores or linear combination scores ("lc") in constrained ordination (cca, rda, capscale). In scores function they are either "vectors" or "factors" (with synonyms "bp" or "cn", resp.). (Description from vegan::envfit help file).}

\item{w}{Weights used in fitting (concerns mainly \code{\link{cca}} and \code{\link{decorana}} results which have nonconstant weights). (Description from vegan::envfit help file).}

\item{na.rm}{Remove rows in ordination scores or \code{cwm} variables with missing values - the whole row of data in both ordination and \code{cwm} is removed in case of any missing values and \code{na.rm = TRUE}.}

\item{test}{Permutation test.}

\item{adjustP}{Logical; should the P-values be adjusted for multiple testing?}

\item{p.adjust.method}{Character string; method of P-value adjustement (check \code{\link{p.adjust.methods}} for available option, default \code{'holm'}).}

\item{p.adjust.n}{Integer; what is the number of comparisons used to adjust the P-values? Default (\code{p.adjust.n = NULL}) means that the number equals to the number of conducted tests (= number of cwm variables).}

\item{...}{Parameters passed to \code{\link{scores}}.}
}
\description{
The function fits community weighted mean of species attributes (e.g. traits), calculated for individual samples, onto an ordination, using multiple regression. The relationship is tested by permutation test (defaults to modified permutation test sensu Zeleny & Schaffers 2012).
}
\details{
The function is derived from the function \code{\link{envfit}} in \code{library (vegan)}, which fits the environmental vectors or factors onto an ordination and tests their significance using permutation test. The modification is in the permutation schema used to test the fit significance: while the original \code{vegan::\link{envfit}} function builds the null distribution of test statistic (r2) by permuting the values of environmental vector (analogy of \code{standard} or \code{rowbased} test), the function \code{weimea::envfit_cwm} additionally offers also \code{modified} or \code{rowbased} test, based on permuting species attributes prior to calculation of \code{cwm}, and \code{max} test combining results of \code{rowbased} and \code{colbased} as \code{P_max = max (P_row, P_col)}. While the null hypothesis of \code{standard} (\code{rowbased}) permutation test that there is "no statistical relationship between mean indicator values and sample ordination scores" is easy to reject in case that CWM's are numerically derived from the same species composition matrix as sample ordination scores (this applies even if some kind of transformation, e.g. log or presence-absence, was applied on the species data before calculating either ordination or CWM), since such relationship is spurious. The null hypothesis of the alternative \code{modified} (\code{rowbased}) permutation test is that there is "no statistical relationship between information in species attributes and sample ordination scores", which is not spurious if species attributes (\code{traits}) are extrinsic to species composition (i.e. not numerically derived from it). The third, \code{max} test, combines these two results together, and is probably not meaningful in the context of relating CWM of species attributes and sample ordination scores, because it includes the \code{standard} (\code{rowbased}) test which has no meaningful null hypothesis (as explained above).

The permutation test requires information about the original vegetation matrix used to calculate the weighted mean of species indicator values, and also original species attributes, which is stored in the object of \code{cwm} class as \code{'com'} and \code{'traits'} attributes. If the argument \code{cwm} is not the object of \code{cwm} class, the function will calculate standard \code{vegan::envfit} and return warning message.

The generic functions \code{print}, \code{plot} and \code{scores} for the class envfit can be used to print the results, plot the vectors of CWM onto an ordination diagram and to return the fitting results. Check \code{?envfit} for details which arguments can be used in these functions.
}
