% Generated by roxygen2 (4.0.1): do not edit by hand
\name{envfit_wm}
\alias{envfit_wm}
\alias{vectorfit_wm}
\title{Regression of ordination scores with weighted mean of species attributes, tested by modified permutation test}
\usage{
envfit_wm(ord, env, permutations = 999, choices = c(1, 2),
  display = "sites", w = weights(ord), na.rm = FALSE, ...)

vectorfit_wm(X, P, permutations, w, ...)
}
\arguments{
\item{ord}{an ordination object of class \code{cca} created by some of the \code{vegan} functions (e.g. \code{\link{rda}}, \code{\link{cca}} or \code{\link{decorana}}) or other type of object from which the ordination scores can be extracted (including an object of the class \code{matrix} or \code{data.frame} with columns of ordination scores).}

\item{env}{object of the class \code{wm} with weighted means of species attributes}
}
\value{
Object of the \code{vegan} class \code{\link{envfit}}
}
\description{
The function fits weighted mean of species attributes (e.g. mean Ellenberg indicator values), calculated for samples, onto an ordination. It employs modified permutation test, based on a null model randomizing species attributes prior to calculation of their sample mean weighted by their abundance. This test is invented to correct the bias in the relationship between mean species attributes (e.g. mean Ellenberg indicator values) and sample scores along ordination axes introduced by the fact that weighted mean of species attributes inherits compositional similarity among samples.
}
\examples{
library (vegan)
data (vltava.spe)
data (vltava.ell)
vltava.mean.eiv <- wm (sitspe = vltava.spe, speatt = vltava.ell)
dca <- decorana (vltava.spe)
# Calculate the fit of mean Ellenberg indicator values with significances based on the original permutation model:
fit.orig <- envfit (ord = dca, env = vltava.mean.eiv, permutations = 499)
# Calculate the fit of mean Ellenberg indicator values with modified permutation model (this expects that you defined the envfit.wm function):
fit.modif <- envfit_wm (ord = dca, env = vltava.mean.eiv, permutations = 499)
# Plot DCA ordination diagram with passively projected Ellenberg indicator values:
plot (dca, display = 'sites', type = 'p')
plot (fit.orig, p.max = .05)
plot (fit.modif, p.max = 0.05, col = 'red')
}
\author{
David Zeleny (zeleny.david@gmail.com); the script is almost entirely based on the original functions envfit and vectorfit from \code{vegan} package, written by Jari Oksanen.
}
\references{
Zeleny D. & Schaffers A.P. (2012): Too good to be true: pitfalls of using mean Ellenberg indicator values in vegetation analyses. Journal of Vegetation Science, 23: 419-431
}
\seealso{
\code{plot.envfit} for plotting the fitted vectors onto ordination diagram, \code{scores.envfit}, \code{print.envfit}
}

