% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ia.R
\name{ia}
\alias{ia}
\title{Whittaker's index of association}
\usage{
ia(sitspe)
}
\arguments{
\item{sitspe}{Compositional matrix (samples x species).}
}
\description{
Creates distance matrix based on Whittaker's index of association, which reflects the dissimilarities between weighted means of species attributes calculated for individual samples, which can be attributed to differences in species composition. The metric is Manhattan-type distance calculated on species profiles (i.e. on community matrix with rows standardized to row-totals).
}
\author{
David Zeleny (zeleny.david@gmail.com)
}
\references{
Legendre P. & Legendre L. 2012. Numerical Ecology. 3rd edn. Elsevier, Oxford, UK.
}

