% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mopet.r
\name{mopet}
\alias{mopet}
\alias{print.mopet}
\alias{summary.mopet}
\title{Modified permutation test with weighted mean of species attributes}
\usage{
mopet(M, env, method = c("lm"), cor.coef = c("pearson"),
  dependence = "M ~ env", permutations = 499, test = "modified",
  parallel = NULL)

\method{print}{mopet}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{mopet}(object, ...)
}
\arguments{
\item{M}{An object of the class \code{wm}}

\item{env}{Vector or matrix with variables. See details.}

\item{method}{Statistical method used to analyse the relationship between M (of class \code{wm} and env), partial match to \code{'lm'}, \code{'aov'}, \code{'cor'}, \code{'kruskal'}, \code{'slope'} and \code{'fourthcorner'}.}

\item{cor.coef}{Correlation coefficient in case of \code{method = 'cor'}. Partial match to 'pearson', 'spearman' and 'kendal'.}

\item{dependence}{Should M be dependent variable and env independent (\code{'M ~ env'}), or opposite? Applicable only for \code{method = 'lm'}. Partial match to \code{'M ~ env'} and \code{'env ~ M'}, so to write \code{dep = 'M'} is enough.}

\item{permutations}{Number of permutations.}

\item{test}{Vector of character values. Which test should be conducted? Partial match of vector with \code{'standard'}, \code{'modified'}, \code{'sequential'} and \code{'all'}.}

\item{parallel}{NULL (default) or integer number. Number of cores for parallel calculation of modified permutation test. Maximum number of cores should correspond to number of available cores on the processor.}

\item{digits}{number of digits reported by \code{print} method on object of \code{"mopet"} class (default is 3).}

\item{...}{Other arguments for \code{print} or \code{summary} functions (not implemented).}

\item{object,x}{object of the class \code{"mopet"}, generated by function \code{mopet}.}
}
\value{
Function \code{mopet} returns list of the class \code{"mopet"} (with \code{print} and \code{summary} methods), which contains the following items:
\itemize{
 \item \code{real.summaries} summary of the method
 \item \code{coefs} model coefficients
 \item \code{stat} test statistic
 \item \code{orig.P} P-values from the original (parametric) test
 \item \code{perm.P} P-values from the not-modified permutation test (permuted are the whole rows of matrix M)
 \item \code{modif.P} P-values from the modified permutation test (permuted are species attributes in object M)
 \item \code{seq.P} P-values from sequential test (\code{max (perm.P, modif.P)})
 \item \code{permutations} number of permutations
 \item \code{method} method of calculation
 \item \code{dependence} dependence (in case of \code{method = 'lm'})
 }
}
\description{
Function performing modified permutation test to calculate significance of relationship between weighted mean of species attributes and other variables.
}
\details{
Currently implemented statistical methods are correlation (\code{'cor'}), linear regression (\code{method = 'lm'}), ANOVA (\code{'aov'}) and Kruskal-Wallis test (\code{'kruskal'}).

Argument \code{env} can be vector or matrix with one column. Only in case of linear regression (\code{method = 'lm'}) is possible to use matrix with several variables, which will be all used as independent variables in the model. For ANOVA and Kruskal-Wallis test, make sure that 'env' is \code{factor} (warning will be returned if this is not the case, but the calculation will be conducted).

Difference between \code{method = 'lm'} and \code{'aov'} is in the format of summary tables, returned by \code{summary.mopet} function. In case of 'aov', this summary is expressed in the traditional language of ANOVA rather than linear models.

Both \code{method = 'lm'} and \code{'slope'} are based on linear regression and calculated by function \code{\link{lm}}, but differ by test statistic: while 'lm' is using F value and is testing the strength of the regression (measured by r2), 'slope' is using the slope of the regression line (b). This statistic is added here for comparison with the fourth corner method. While r2 (and r) is influenced by the issue of compositional autocorrelation, slope of regression is not.

Specific issue related to weighted mean is the case of missing species attributes. In current implementation, species with missing species attributes are removed from sample x species matrix prior to permutation of species attributes among species.
}
\examples{
data (vltava)
mean.eiv <- wm (vltava$spe, vltava$ell)
mopet (mean.eiv, vltava$env$pH, perm = 49, test = 'all')
mopet (mean.eiv, vltava$env$pH, perm = 49, test = 'stand', method = 'cor', cor.coef = 'spearm')
summary (mopet (mean.eiv, vltava$group, perm = 49, method = 'krusk'))
}
\seealso{
\code{\link{wm}}
}

