% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.r
\name{randomize}
\alias{randomize}
\alias{randomize.cwm}
\title{Weighted mean calculated from randomized species attributes}
\usage{
randomize(x, ...)

\method{randomize}{cwm}(
  x,
  permutations = 1,
  FUN = function(y) y,
  parallel = NULL,
  library.par = NULL,
  export.cl = NULL,
  ...
)
}
\arguments{
\item{x}{object of the class \code{cwm}}

\item{...}{parameters passed to other functions.}

\item{permutations}{number of randomizations}

\item{FUN}{function to be applied on the column of calculated weighted mean values}

\item{parallel}{integer or NULL (default). If integer, calculation will be conducted in parallel on number of cores given by \code{parallel}}

\item{library.par}{character vector with libraries needed for application of function defined by \code{FUN} argument (to be exported into parallel process). Not necessary if the functions in \code{FUN} are stated explicitly (e.g. \code{vegan:::cca})}

\item{export.cl}{??}
}
\description{
Function applied on the object of class \code{cwm}, which returns values of weighted mean calculated from randomized species attributes.
}
\author{
David Zeleny (zeleny.david@gmail.com)
}
