% Generated by roxygen2 (4.0.1): do not edit by hand
\name{range.wm}
\alias{range.wm}
\title{Changes the range of species attributes}
\usage{
\method{range}{wm}(object, max = 9, recalculate = T)
}
\arguments{
\item{object}{object of the class \code{wm}}

\item{max}{upper bound of the rescaled values}

\item{recalculate}{should the mean values in \code{object} be recalculated?}
}
\description{
Changes the range of species attributes
}

