% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeWM.r
\name{rangeWM}
\alias{rangeWM}
\title{Changes the range of species attributes}
\usage{
rangeWM(object, max = 9, recalculate = T)
}
\arguments{
\item{object}{object of the class \code{wm}}

\item{max}{upper bound of the rescaled values}

\item{recalculate}{should the mean values in \code{object} be recalculated?}
}
\description{
Changes the range of species attributes
}

