% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simul.comm.2.r
\name{simul.comm.2}
\alias{sample.comm.2}
\alias{simul.comm.2}
\title{Simulation of ecological data structured by two ecological gradients}
\usage{
simul.comm.2(totS = 300, gr1.length = 5000, gr2.length = 5000,
  niche.type = "random", max.niche.breath = c(gr1.length, gr2.length),
  seed = NULL, plotting = F)

sample.comm.2(simul.comm = simul.comm.2(), Np = 300, sample.x1 = NULL,
  sample.x2 = NULL, no.ind = 100, no.spec = 20, seed = NULL, pa = F,
  based.on = "individuals")
}
\arguments{
\item{totS}{total number of species in simulation}

\item{gr1.length,gr2.length}{length of first and second gradient in units}

\item{niche.type}{shape of species response curves ('random', 'normal', 'skewed')}

\item{max.niche.breath}{vector of the length = 2, with maximum niche breath along first and second gradient (default = \code{c(gr1.length, gr2.length)})}

\item{seed}{set random seed for reproducing always the same result}

\item{plotting}{should the species response curves along simulated gradient be drawned?}

\item{simul.comm}{result of simul.comm.2 function with parameters of individual species response curves}

\item{Np}{number of samples}

\item{sample.x1,sample.x2}{positions of sampling along the first (or second, respectively) gradient (default = NULL, means that random locations are generated)}

\item{no.ind}{mean number of individuals to be drawn from the species pool (if \code{based.on = 'individuals'})}

\item{no.spec}{number of species to be drawn from species pool (if \code{based.on = 'species'})}

\item{pa}{result table will be generated in presence-absence form (default \code{pa = TRUE})}

\item{based.on}{should the sampling of species from species pool be based on \code{'individuals'} or \code{'species'}? (default = \code{'individuals'})}
}
\description{
Simulation of ecological data structured by two ecological gradients
}
\examples{
sc <- sample.comm.2 (simul.comm.2 (gr1.length = 5000, gr2.length = 3000, totS = 300),
 Np = 200, pa = FALSE, seed = 123)
library (vegan)
nmds <- metaMDS (sc$a.mat)
par (mfrow = c(1,2))
ordiplot (nmds, display = 'si', main = 'Environmental gradient 1')
points (nmds, display = 'si', cex = scale (sc$sample.x1, center = FALSE)*2, pch = 21, bg = 'white')
ordiplot (nmds, display = 'si', main = 'Environmental gradient 2')
points (nmds, display = 'si', cex = scale (sc$sample.x2, center = FALSE)*2, pch = 21, bg = 'white')
}
\author{
David Zeleny (zeleny.david@gmail.com) - based on the script of Jason Fridley (Fridley et al. 2007, Appendix S2), modified for two dimensions
}

