% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snc.R
\name{snc}
\alias{snc}
\alias{is.snc}
\alias{[.snc}
\alias{as.matrix.snc}
\alias{summary.snc}
\alias{print.snc}
\alias{extract.snc}
\title{Weighted mean of sample attributes - species niche centroids (SNC)}
\usage{
snc(com, env, wstand = FALSE)

is.snc(object)

\method{[}{snc}(object, i, j)

\method{as.matrix}{snc}(x, ...)

\method{summary}{snc}(object, long = F, ...)

\method{print}{snc}(x, ...)

\method{extract}{snc}(x, what = "env", ...)
}
\arguments{
\item{com}{Matrix or data.frame with species composition matrix (samples x species)}

\item{env}{Vector, matrix or data.frame with env. variables (samples x env). This can be \code{numeric} or \code{factor}.}

\item{wstand}{Logical. Should be the values in \code{env} weighted-standardized prior to calculation of SNC? Weights are row sums in species composition matrix \code{com}.}

\item{object, x}{Object of the class \code{snc}}

\item{i, j}{Subscripts of the matrix of the class "snc" (rows and columns).}

\item{...}{Other arguments passed into \code{as.matrix}, \code{summary} or \code{print} function. Currently not supported.}

\item{long}{should summary return long output? (TRUE vs FALSE)}

\item{what}{Attributes extracted from the object of class \code{snc}; either \code{env} for the matrix of sample attributes (samples x env), or \code{com} for matrix of species composition (samples x species).}
}
\value{
Object of class \code{"snc"}, which has \code{print}, \code{summary}, \code{as.matrix} and \code{'[.'} methods. Object of \code{"snc"} class contains the matrix of calculated weighted means of sample attributes (species niche centroids) for each species (species x snc) and three attributes: \code{com}, species x sample matrix from which the weighted mean was calculated, \code{env}, samples x env. variables matrix with sample attributes, and \code{wstand}, logical value specifying whether env.variables were weighted-standardized prior to calculation of SNC (weights are row sums of \code{com}). All weighted means of sample attributes must be based on the same species x sample matrix with the same number of species.
}
\description{
NEEDS TO REWRITE!!!
Function \code{snc} calculates weighted mean of sample attributes, using matrix of species composition (\code{com}) and one or several sample attributes (\code{env}).
}
\details{
Function \code{[.snc]} is for extracting specified rows and columns from matrix of class \code{snc}. As a side effect, resulting object will have concatenated the \code{com} and \code{env} attributes to match the dimension of the resulting matrix. This function is only for extracting the parts of \code{snc} object, not for replacing; attempt to replace will work, but will break the inner structure.

Generic function \code{extract} with argument \code{what} applied on object of class \code{"snc"} extracts the original species composition matrix or sample attribute matrix, respectively, which were used to calculate weighted means.
}
\examples{
# Calculation of weighted mean of species Ellenberg indicator values using dataset Vltava
data (vltava)
mean.env <- snc(com = vltava$spe, env = vltava$env[, c('ELEVATION', 'pH')])

summary (mean.env)

# Extracting values from the object of \code{snc} class
mean.env[,1]
mean.env[1:10, 1]
}
\author{
David Zeleny (zeleny.david@gmail.com)
}
