% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.LR.r
\name{test.LR}
\alias{coef.testLR}
\alias{print.testLR}
\alias{summary.testLR}
\alias{test.LR}
\alias{test.LR.0}
\title{Test of the link between sample attributes and matrix of species composition (i.e. whether \bold{L} is linked to \bold{R})}
\usage{
test.LR(M, env, type = "cor", cor.coef = c("pearson", "kendall",
  "spearman"), exact = FALSE, alpha = 0.001, sqrt = F, perm = 499)

test.LR.0(M, env, type = "cor", cor.coef = c("pearson"), exact = FALSE,
  alpha = 0.001, sqrt = F, perm = 199)

\method{print}{testLR}(x, digits = 3, ...)

\method{summary}{testLR}(object, ...)

\method{coef}{testLR}(object, ...)
}
\arguments{
\item{M}{Object of the 'wm' class. Matrix with weighted means of species attributes.}

\item{env}{Matrix with environmental variables.}

\item{type}{Currently not implemented. In the future, other types of the test (apart to the one based on db-RDA) should be available.}

\item{cor.coef}{If \code{type = "cor"}: which correlation coefficient should be used? Partial match to \code{c("pearson", "spearman", "kendal")}.}

\item{exact}{If \code{type = "cor"} and \code{cor.coef = "spearman"} or \code{cor.coef = 'kendal'} - a logical indicating whether and exact p-value should be computed.}

\item{alpha}{Target Type I error rate for Monte Carlo permutation tests (influences number of permutations).}

\item{sqrt}{Logical value, default FALSE. Should the distance matrix based on Whittaker's index of association be square-rooted to become Euclidean? See Details.}

\item{perm}{Number of permutations for \code{type = "cor"}. Default = 199.}

\item{x,object}{Object of the class "testLR"}

\item{digits}{Number of digits reported for parameters in summary output.}

\item{...}{Other arguments passed into \code{print}, \code{summary} or \code{coef} functions (not implemented yet).}
}
\value{
Object of the class "testLR", list of lists, each node containing two parts: results of dbRDA analysis (calculated by \code{\link{capscale}} function from \code{vegan}) and results of Monte Carlo permutation test (calculated by \code{\link{anova.cca}} function, also from \code{vegan}).
}
\description{
The function testing the link between sample attributes and species composition of the matrix, from which weighted means are calculated. The test is based either on db-RDA (distance-based redundancy analysis), or Moran's I. Significant relationship is considered as an argument to use modified permutation test instead of the standard permutation test for testing the relationship between weighted mean of species attributes and sample attributes.
}
\details{
In case of dbRDA, the matrix of intersample distances is calculated using Whittaker's index of association (\code{\link{ia}}) and significance of the variation explained by sample attributes (R2) is tested by Monte Carlo permutation test. In case of Moran's I, the test is examining wheather the sample attributes variable is compositionally autocorrelated, i.e. whether the Moran's I calculated on this variable using as weighted inverted dissimilarities between sample's species composition is significant.

Whittaker's index of association (calculated as Manhattan type distance on species profiles) is metric, but not Euclidean, and in PCoA (on which dbRDA is based) it can produce negative eigenvalues. After square root transformation, the index becomes both metric and Euclidean.

Variation explained by given environmental variable (R2) can differ between individual weighted means calculated from the same species composition matrix. This happens when different species attributes have different values missing; explained variation is calculated only from those columns (species) of compositional matrix L, which have assigned value for given species attribute.
}
\examples{
data (vltava)
test.LR (M = wm (vltava$spe, vltava$ell), vltava$env[,'pH', drop = FALSE], type = 'cor')
}

