% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.MR.r
\name{test.MR}
\alias{coef.testMR}
\alias{print.testMR}
\alias{test.MR}
\title{Testing the relationship between weighted-mean of species attributes with sample attributes}
\usage{
test.MR(M, env, method = c("cor"), cor.coef = c("pearson"),
  dependence = "M ~ env", perm = 499, testLR.perm = NULL,
  test = "twostep", fc.test = 6, parallel = NULL, testLR.P = 0.05)

\method{print}{testMR}(x, digits = max(3, getOption("digits") - 3), ...)

\method{coef}{testMR}(object, ...)
}
\arguments{
\item{M}{An object of the class \code{wm}}

\item{env}{Vector or matrix with variables. See details.}

\item{method}{Statistical method used to analyse the relationship between M (of class \code{wm}) and env (sample attributes); partial match to \code{'lm'}, \code{'cor'} and \code{'fourthcorner'}.}

\item{cor.coef}{Correlation coefficient for \code{method = 'cor'}. Partial match to \code{'pearson'} and \code{'spearman'}.}

\item{dependence}{Should \code{M} be dependent variable and \code{env} independent (\code{'M ~ env'}), or opposite? Applicable only for \code{method = 'lm'}. Partial match to \code{'M ~ env'} and \code{'env ~ M'}.}

\item{perm}{Number of permutations.}

\item{testLR.perm}{Relevant if \code{test = "twostep"} (default): number of permutations for test of relationship between \bold{L} and \bold{R} matrices.}

\item{test}{Vector of character values. Which test should be conducted? Partial match of vector with \code{'standard'}, \code{'modified'}, \code{'twostep'}, \code{'sequential'} and \code{'all'}. See \code{Details}.}

\item{fc.test}{Test for the fourthcorner method. \code{fc.test = c(2,4,6)}.}

\item{parallel}{NULL (default) or integer number. Number of cores for parallel calculation of modified permutation test. Maximum number of cores should correspond to number of available cores on the processor.}

\item{testLR.P}{Relevant if \code{test = "twostep"} (default): significance value at whitch the relationship between \bold{L} and \bold{R} matrices are significant.}

\item{x,object}{object of the class \code{"wm"}, generated by function \code{wm}.}

\item{digits}{number of digits reported by \code{print} method on object of \code{"wm"} class (default is 3).}

\item{...}{Other arguments for \code{print}, \code{summary} or \code{coef} functions (not implemented yet).}
}
\value{
Function \code{wm} returns list of the class \code{"wm"} (with \code{print} and \code{summary} methods), which contains the following items:
\itemize{
 \item \code{real.summaries} summary of the method
 \item \code{coefs} model coefficients
 \item \code{stat} test statistic
 \item \code{orig.P} P-values from the original (parametric) test
 \item \code{perm.P} P-values from the not-modified permutation test (permuted are the whole rows of matrix M)
 \item \code{modif.P} P-values from the modified permutation test (permuted are species attributes in object M)
 \item \code{seq.P} P-values from sequential test (\code{max (perm.P, modif.P)})
 \item \code{perm} number of permutations
 \item \code{method} method of calculation
 \item \code{dependence} dependence (in case of \code{method = 'lm'})
 }
}
\description{
Function performing standard, modified, two-step and sequential test to calculate significance of relationship between weighted-mean of species attributes and sample attributes.
}
\details{
Currently implemented statistical method is correlation (\code{'cor'}).

Argument \code{env} can be vector or matrix with one column. Only in case of linear regression (\code{method = 'lm'}) is possible to use matrix with several variables, which will be all used as independent variables in the model. For ANOVA and Kruskal-Wallis test, make sure that 'env' is \code{factor} (warning will be returned if this is not the case, but the calculation will be conducted).

Difference between \code{method = 'lm'} and \code{'aov'} is in the format of summary tables, returned by \code{summary.wm} function. In case of 'aov', this summary is expressed in the traditional language of ANOVA rather than linear models.

Both \code{method = 'lm'} and \code{'slope'} are based on linear regression and calculated by function \code{\link{lm}}, but differ by test statistic: while 'lm' is using F value and is testing the strength of the regression (measured by r2), 'slope' is using the slope of the regression line (b). This statistic is added here for comparison with the fourth corner method. While r2 (and r) is influenced by the issue of compositional autocorrelation, slope of regression is not.

Specific issue related to weighted mean is the case of missing species attributes. In current implementation, species with missing species attributes are removed from sample x species matrix prior to permutation of species attributes among species.
}
\examples{
data (vltava)
M <- wm (sitspe = vltava$herbs$spe, speatt = vltava$herbs$traits)
re <- test.MR (M = M, env = vltava$env[,c('pH', 'COVERE32')])
}
\seealso{
\code{\link{wm}}
}

