% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fourth.R
\name{test_fourth}
\alias{test_fourth}
\alias{print.testFOURTH}
\alias{coef.testFOURTH}
\title{Testing the relationship species attributes and sample attributes by fourth corner method}
\usage{
test_fourth(env, com = NULL, traits = NULL, cwm = NULL, perm = 499,
  test = "max", parallel = NULL, p.adjust.method = p.adjust.methods,
  adjustP = FALSE)

\method{print}{testFOURTH}(x, digits = max(3, getOption("digits") - 3),
  missing.summary = TRUE, ...)

\method{coef}{testFOURTH}(object, ...)
}
\arguments{
\item{env}{Vector, matrix or data frame with sample attributes (environemntal variables).}

\item{com}{Matrix of data frame with species composition data.}

\item{traits}{Vector, matrix or data frame with species attributes (traits)}

\item{cwm}{An object of the class \code{cwm}.}

\item{perm}{Number of permutations.}

\item{test}{Vector of character values. Which test should be conducted? Partial match to \code{'rowbased'} row-based permutation test (called Model 2 in the Legendre et al. 1997 and Dray & Legendre 2008), \code{'colbased'} for column-based permutation test (Model 4), and \code{'max'} for max test (selecting the higher from \code{rowbased} and \code{colbased} result, Model 6). \code{'all'} includes all three tests. See \code{Details}.}

\item{parallel}{NULL (default) or integer number. Number of cores for parallel calculation of modified permutation test. Maximum number of cores should correspond to number of available cores on the processor.}

\item{p.adjust.method}{A string indicating the method of P-value adjustement, see \code{\link{p.adjust.methods}} for possible choices.}

\item{adjustP}{Logical, default FALSE. Should be the P-values adjusted? If \code{adjustP = TRUE}, the last column in the results is adjusted by method selected in \code{p.adjust.method}.}

\item{x, object}{object of the class \code{"cwm"}, generated by function \code{cwm}.}

\item{digits}{number of digits reported by \code{print} method on object of \code{"cwm"} class (default is 3).}

\item{missing.summary}{Logical; should be the summary of values missing in \code{env}, \code{cwm} and \code{traits} be printed along to the result of \code{test_cwm} analysis? Default is \code{TRUE}.}

\item{...}{Other arguments for \code{print}, \code{summary} or \code{coef} functions (not implemented yet).}
}
\value{
Function \code{test_fourth} returns the list of the class \code{"testFOURTH"} (with \code{print} and \code{summary} methods), which contains the following items:
\itemize{
 \item \code{call} Function call.
 \item \code{out} Data frame with results of calculation; first two columns contain the original fourth corner metric (\code{r_fc}) and its Chessel transformation (\code{r_ch}); following columns contain some of the following results of permutation tests: \code{P.row}, \code{P.col} and \code{P.max}. If \code{adjustP = TRUE}, the last column contains P values adjusted by \code{p.adjust} function.
 \item \code{miss} List with missing values of \code{env}, \code{traits} or \code{cwm} (the last provided for compatibility with \code{test_cwm} function and refers to samples which, after removing missing \code{traits} values, do not contain any species.)
 \item \code{param} Setting of input parameters (\code{perm} and \code{test}).
 }
}
\description{
Function calculating fourth corner analysis between sample attributes (environmental variables) and species attributes (traits) and testing it by row based, column based or max permutation test.
}
\details{
Environmental variables, species composition and traits can be provided in two ways: either each object separately, using arguments \code{env}, \code{com} and \code{traits}, or using only \code{env} and \code{cwm}; in the later case, the function \code{extract} is used to extract the \code{com} and \code{traits} from the \code{cwm} object.
Specific issue related to weighted mean is the case of missing species attributes. In current implementation, species with missing species attributes are removed from sample x species matrix prior to permutation of species attributes among species.
}
\examples{
data (vltava)
vltava.env <- vltava$env[,c('pH', 'COVERE32')]
vltava.com <- vltava$herbs$spe
vltava.traits <- vltava$herbs$traits
fc <- test_fourth (env = vltava.env, com = vltava.com, traits = vltava.traits)
}
\seealso{
\code{\link{cwm}} \code{\link{test_cwm}}
}
