% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wm.r
\name{wm}
\alias{[.wm}
\alias{as.matrix.wm}
\alias{is.wm}
\alias{print.wm}
\alias{summary.wm}
\alias{wm}
\alias{wmC}
\title{Weighted mean of species attributes}
\usage{
wm(sitspe, speatt)

wmC(sitspe, speatt)

is.wm(object)

\method{[}{wm}(object, i, j, drop = F)

\method{as.matrix}{wm}(x, ...)

\method{summary}{wm}(object, long = F, ...)

\method{print}{wm}(x, ...)
}
\arguments{
\item{sitspe}{Matrix or data.frame with community data matrix (sites x species)}

\item{speatt}{Vector, matrix or data.frame with species attributes (species x attributes)}

\item{object,x}{Object of the class \code{wm}}

\item{i,j}{Subscripts of the matrix of the class "Wm" (rows and columns).}

\item{drop}{In function \code{[.wm]} if \code{TRUE} coerces the result to the lowest possible dimension (i.e. vector if matrix has only one column). Currently not implemented, change into \code{TRUE} will have no effect.}

\item{...}{Other arguments passed into \code{as.matrix}, \code{summary} or \code{print} function. Currently not supported.}

\item{long}{should summary return long output? (TRUE vs FALSE)}
}
\value{
Object of class \code{"wm"}, which has \code{print}, \code{summary}, \code{as.matrix} and \code{'[.'} methods. Object of \code{"wm"} class contains the matrix of calculated weighted means of species attributes for each sample (sample x weighted mean) and two attributes: \code{sitspe} species x sample matrix from which the weighted mean was calculated, and \code{speatt} species x attributes matrix with species attributes. All weighted means of species attributes must be based on the same species x sample matrix with the same number of samples.
}
\description{
Function \code{wm} calculates weighted mean of species attributes, using matrices of species composition and species attributes. Other functions are for handling the objects of class \code{wm}.
}
\details{
Function \code{[.wm]} is for extracting specified rows and columns from matrix of class \code{wm}. As a side effect, resulting object will have concatenated the \code{sitspe} and \code{speatt} attributes to match the dimension of the resulting matrix. This function is only for extracting the parts of \code{wm} object, not for replacing! (attempt to replace will work, but will break the inner structure).

Generic functions \code{sitspe} and \code{speatt} applied on object of class \code{"wm"} extracts the original species composition matrix and species attribute matrix, respectively, which were used to calculate weighted means.
}
\examples{
# Calculation of weighted mean of species Ellenberg indicator values using dataset Vltava
data (vltava)
mean.eiv <- wm (sitspe = vltava$spe, speatt = vltava$ell)

summary (mean.eiv)

# Extracting values from the object of \\code{wm} class
mean.eiv[,1]
mean.eiv[1:10, 2:3]
}
\author{
David Zeleny (zeleny.david@gmail.com)
}

