% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fourth.corner.r
\name{fourth.corner}
\alias{fourth.corner}
\title{calculates the fourth-corner statistic (according to Legendre et al. 1997, compiled according to Peres-Neto et al. 2012)}
\usage{
fourth.corner(R, L, Q)
}
\arguments{
\item{R}{Matrix of sample attributes.}

\item{L}{Matrix of species composition.}

\item{Q}{Matrix of species attributes}

\item{fc.test}{Test to be chosen for fourthcorner analysis.}

\item{perm}{Number of permutations.}
}
\description{
calculates the fourth-corner statistic (according to Legendre et al. 1997, compiled according to Peres-Neto et al. 2012)
}

