% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ia}
\alias{ia}
\title{Whittaker's index of association}
\usage{
ia(sitspe)
}
\arguments{
\item{sitspe}{Compositional matrix (samples x species).}
}
\description{
Creates distance matrix based on Whittaker's index of association, which is reflects the distances between weighted means of species attributes calculated for individual samples. The metric is baselly Manhattan-type distance calculated on species profiles (i.e. on community matrix in which individual rows are standardized to row-totals).
}
\author{
David Zeleny (zeleny.david@gmail.com)
}
\references{
Legendre P. & Legendre L. 2012. Numerical Ecology. 3rd edn. Elsevier, Oxford, UK.
}

