% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomize.r
\name{randomize}
\alias{randomize}
\alias{randomize.wm}
\title{Weighted mean calculated from randoxized species attributes}
\usage{
randomize(x, ...)

\method{randomize}{wm}(x, permutations = 1, FUN = function(y) y,
  progress.bar = F, parallel = NULL, library.par = NULL,
  export.cl = NULL, ...)
}
\arguments{
\item{x}{object of the class \code{wm}}

\item{...}{parameters passed to other functions.}

\item{permutations}{number of randomizations}

\item{FUN}{function to be applied on the column of calculated weighted mean values}

\item{progress.bar}{logical value, should be the progress bar indicating the progress of the calculation launched?}

\item{parallel}{integer or NULL (default). If integer, calculation will be conducted in parallel on number of cores given by \code{parallel}}

\item{library.par}{character vector with libraries needed for application of function defined by \code{FUN} argument (to be exported into parallel process). Not necessary if the functions in \code{FUN} are stated explicitly (e.g. \code{vegan:::cca})}

\item{export.cl}{??}
}
\description{
Function applied on the object of class \code{wm}, which returns values of weighted mean calculated from randomized species attributes.
}
\author{
David Zeleny (zeleny.david@gmail.com)
}

