% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cwm.R
\name{test_cwm}
\alias{test_cwm}
\alias{print.testCWM}
\alias{coef.testCWM}
\alias{plot.testCWM}
\title{Testing the relationship between weighted-mean of species attributes with sample attributes}
\usage{
test_cwm(cwm, env, method = c("cor"), wcor = FALSE, wstand = FALSE,
  wreg = NULL, dependence = "cwm ~ env", perm = 499, test = "max",
  parallel = NULL, p.adjust.method = "holm", adjustP = FALSE)

\method{print}{testCWM}(x, digits = max(3, getOption("digits") - 3),
  missing.summary = FALSE, adjustP = FALSE, ...)

\method{coef}{testCWM}(object, ...)

\method{plot}{testCWM}(x, alpha = 0.05, line = 2.5, cex.lab = 1.5,
  par.mar = c(4, 4, 0.5, 0.5), box.col = c("blue", "red"), box.lwd = 2,
  ...)
}
\arguments{
\item{cwm}{An object of the class \code{cwm}.}

\item{env}{Vector or matrix with variables. See details.}

\item{method}{Statistical method used to analyse the relationship between cwm (of class \code{cwm}) and env (sample attributes); partial match to \code{'cor'}, \code{'lm'} and \code{'aov'}.}

\item{wcor}{Logical; should the correlation be weighted by rowsums of \code{com} (extracted from \code{cwm}?). Default \code{FALSE}.}

\item{wstand}{Logical; should the variables in correlation be first weighted-standardized? Default \code{FALSE}. If \code{wstand = TRUE}, both \code{env} and \code{traits} are weighted standardized prior to calcuation; weights are derived from \code{com} matrix (extracted from \code{cwm}); \code{env} is weighted by rowsums of \code{com}, while \code{traits} are weighted by colsums of \code{com}.}

\item{wreg}{\code{NULL} (default) or numeric vector of weights used in regression (\code{method = 'lm'}). If weights are provided, weighted least squares are calculated instead of ordinary least squares.}

\item{dependence}{Should \code{cwm} be dependent variable and \code{env} independent (\code{'cwm ~ env'}), or opposite? Applicable only for \code{method = 'lm'}. Partial match to \code{'cwm ~ env'} and \code{'env ~ cwm'}.}

\item{perm}{Number of permutations.}

\item{test}{Vector of character values. Which test should be conducted? Partial match to \code{'standard'} or \code{'rowbased'} for standard (row based) permutation test, \code{'modified'} or \code{'colbased'} for modified (column based) permutation test, \code{'max'} for max test (selecting the higher from \code{rowbased} and \code{colbased} result), and \code{'all'} including all three tests. See \code{Details}.}

\item{parallel}{NULL (default) or integer number. Number of cores for parallel calculation of modified permutation test. Maximum number of cores should correspond to number of available cores on the processor.}

\item{p.adjust.method}{A string indicating the method of P-value adjustement, see \code{\link{p.adjust.methods}} for possible choices.}

\item{adjustP}{Logical, default FALSE. Should be the P-values adjusted? If \code{adjustP = TRUE}, the last column in the results is adjusted by method selected in \code{p.adjust.method}.}

\item{x, object}{object of the class \code{"cwm"}, generated by function \code{cwm}.}

\item{digits}{number of digits reported by \code{print} method on object of \code{"cwm"} class (default is 3).}

\item{missing.summary}{Logical; should be the summary of values missing in \code{env}, \code{cwm} and \code{traits} be printed along to the result of \code{test_cwm} analysis? Default is \code{TRUE}.}

\item{...}{Other arguments for \code{print}, \code{summary}, \code{coef} or \code{plot} functions (some not implemented yet).}

\item{alpha, line, cex.lab, par.mar, box.col, box.lwd}{Graphical parameters for \code{plot} function.}
}
\value{
Function \code{cwm} returns list of the class \code{"cwm"} (with \code{print} and \code{summary} methods), which contains the following components:
\itemize{
 \item \code{call} Call to the function.
 \item \code{out} Matrix with analysis results (coefficients, statistics, P-values).
 \item \code{miss} Matrix with counts of missing values in \code{env}, \code{cwm} and \code{traits}.
 \item \code{param} List with the setting of the function parameters (arguments).
}
}
\description{
Function calculating relationship between weighted-mean of species attributes and sample attributes and performing standard (row based), modified (column based), or max test of significance.
}
\details{
Currently implemented statistical methods: \code{'cor'}. Plan to implement also: \code{'lm'} and \code{'aov'}. For fourth corner, please use \code{\link{test_fourth}}.

Argument \code{env} can be vector or matrix with one column. Only in the case of linear regression (\code{method = 'lm'}) it is possible to use matrix with several variables, which will all be used as independent variables in the model. For ANOVA and Kruskal-Wallis test, make sure that 'env' is \code{factor} (warning will be returned if this is not the case, but the calculation will be conducted). 

Difference between \code{method = 'lm'} and \code{'aov'} is in the format of summary tables, returned by \code{summary.cwm} function. In case of 'aov', this summary is expressed in the traditional language of ANOVA rather than linear models.

Both \code{method = 'lm'} and \code{'slope'} are based on linear regression and calculated by function \code{\link{lm}}, but differ by test statistic: while 'lm' is using F value and is testing the strength of the regression (measured by r2), 'slope' is using the slope of the regression line (b). This statistic is added here for comparison with the fourth corner method.

Specific issue related to weighted mean is the case of missing species attributes. In current implementation, species with missing species attributes are removed from sample x species matrix prior to permutation of species attributes among species.
}
\examples{
data (vltava)
CWM <- cwm (com = vltava$herbs$spe, traits = vltava$herbs$traits)
re <- test_cwm (cwm = CWM, env = vltava$env[,c('pH', 'COVERE32')])
re
plot (re)
}
\seealso{
\code{\link{cwm}}, \code{\link{snc}}
}
