% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weimea.R
\docType{package}
\name{weimea}
\alias{weimea}
\title{weimea - package for calculating and testing weighted-means of species attributes with sample attributes}
\description{
Calculates CWM and SNC and tests its relationship to environmental variables
}
\author{
David Zeleny
}
