% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weimea.R
\docType{package}
\name{weimea}
\alias{weimea}
\alias{weimea-package}
\title{Weimea - package for calculating and testing weighted-means of species attributes with sample attributes}
\description{
Weimea - package for calculating and testing weighted-means of species attributes with sample attributes
}
\author{
David Zeleny
}

