% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wm}
\alias{[.wm}
\alias{as.matrix.wm}
\alias{is.wm}
\alias{print.wm}
\alias{summary.wm}
\alias{wm}
\title{Weighted mean of species attributes}
\usage{
wm(sitspe, speatt)

is.wm(object)

\method{[}{wm}(M, i, j, drop = F)

\method{as.matrix}{wm}(M)

\method{summary}{wm}(M, long = F)

\method{print}{wm}(M)
}
\arguments{
\item{sitspe}{Matrix or data.frame with community data matrix (sites x species)}

\item{speatt}{Vector, matrix or data.frame with species attributes (species x attributes)}

\item{object}{any R object}

\item{M}{Object of the class \code{wm}}

\item{i,j}{Subscripts of the matrix of the class "Wm" (rows and columns).}

\item{drop}{In function \code{[.wm]} if \code{TRUE} coerces the result to the lowest possible dimension (i.e. vector if matrix has only one column). Currently not implemented, change into \code{TRUE} will have no effect.}

\item{long}{should summary return long output? (TRUE vs FALSE)}
}
\value{
Object of class \code{\link{wm}}, which contains the matrix of calculated weighted means of species attributes for each sample (sample x weighted mean) and two attributes: \code{sitspe} species x sample matrix from which the weighted mean was calculated, and \code{speatt} species x attributes matrix with species attributes. All weighted means of species attributes must be based on the same species x sample matrix with the same number of samples.
}
\description{
Function \code{wm} calculates weighted mean of species attributes, using matrices of species composition and species attributes. Other functions are for handling the objects of class \code{wm}.
}
\details{
Function \code{[.wm]} is for extracting specified rows and columns from matrix of class \code{wm}. As a side effect, resulting object will have concatenated the \code{sitspe} and \code{speatt} attributes to match the dimension of the resulting matrix. This function is only for extracting the parts of \code{wm} object, not for replacing! (attempt to replace will work, but will break the inner structure).
}
\examples{
# Calculation of weighted mean of species Ellenberg indicator values using dataset Vltava
data (vltava)
mean.eiv <- wm (sitspe = vltava$spe, speatt = vltava$ell)

# Calculation of community weighted mean of species functional traits using dataset aravo from ade4 library
data (aravo, package = 'ade4')
cwm <- wm (sitspe = aravo$spe, speatt = aravo$traits)

summary (cwm) # brief summary

# Extracting values from the object of \\code{wm} class
cwm[,1]
cwm[1:10, 2:3]
}
\author{
David Zeleny (zeleny.david@gmail.com)
}

